+++ client/js/scorm/components/scorm/settings.jsx
@@ -2,7 +2,7 @@ import React             from 'react';
 import HoverButton       from '../common/hover_button';
 import Defines           from '../../../defines';
 import CommonSvg         from '../../../common_components/common_svg';
-import ConfirmDelete from './confirm_delete';
+import ConfirmDelete from '../../../common_components/confirm_delete';
 
 
 export default function settings(props) {
